refactor(parsers,stores) move enums from stores to parsers
